hlsenc: Fix the return value accumulation in append_single_file
Both the read_byte variable (which is accumulated into
append_single_file) and the return value are int64_t;
give the ret variable the right corresponding type too.

+++ libavformat/hlsenc.c
@@ -2380,7 +2380,7 @@ static int hls_init_file_resend(AVFormatContext *s, VariantStream *vs)
 
 static int64_t append_single_file(AVFormatContext *s, VariantStream *vs)
 {
-    int ret = 0;
+    int64_t ret = 0;
     int64_t read_byte = 0;
     int64_t total_size = 0;
     char *filename = NULL;
